+++ packages/app/hivecommand-api/src/gqty/schema.generated.ts
@@ -86,6 +86,7 @@ export interface CommandDevicePeripheralInput {
 }
 
 export interface CommandDevicePluginConfigurationInput {
+  id?: InputMaybe<Scalars["ID"]>;
   key?: InputMaybe<Scalars["String"]>;
   value?: InputMaybe<Scalars["String"]>;
 }
@@ -143,6 +144,14 @@ export interface CommandPeripheralProductInput {
   name?: InputMaybe<Scalars["String"]>;
   port?: InputMaybe<Scalars["String"]>;
   vendorId?: InputMaybe<Scalars["String"]>;
+}
+
+export interface CommandProgramDataDeviceInterlockInput {
+  assertion?: InputMaybe<CommandAssertionInput>;
+  comparator?: InputMaybe<Scalars["String"]>;
+  deviceKey?: InputMaybe<Scalars["String"]>;
+  inputDevice?: InputMaybe<Scalars["String"]>;
+  inputDeviceKey?: InputMaybe<Scalars["String"]>;
 }
 
 export interface CommandProgramDeviceCalibrationInput {
@@ -301,6 +310,16 @@ export const generatedSchema = {
     value: { __type: "String" },
     variable: { __type: "String" },
   },
+  CommandDataInterlock: {
+    __typename: { __type: "String!" },
+    assertion: { __type: "CommandAssertion" },
+    comparator: { __type: "String" },
+    device: { __type: "CommandProgramDevicePlaceholder" },
+    deviceKey: { __type: "CommandProgramDeviceState" },
+    id: { __type: "ID!" },
+    inputDevice: { __type: "CommandProgramDevicePlaceholder" },
+    inputDeviceKey: { __type: "CommandProgramDeviceState" },
+  },
   CommandDevice: {
     __typename: { __type: "String!" },
     activeProgram: { __type: "CommandProgram" },
@@ -316,6 +335,7 @@ export const generatedSchema = {
     organisation: { __type: "HiveOrganisation" },
     peripherals: { __type: "[CommandDevicePeripheral]" },
     reports: { __type: "[CommandDeviceReport]" },
+    setpoints: { __type: "[CommandDeviceSetpointCalibration]" },
     waitingForActions: { __type: "[CommandProgramAction]" },
   },
   CommandDeviceInput: {
@@ -367,12 +387,19 @@ export const generatedSchema = {
   },
   CommandDevicePlugin: {
     __typename: { __type: "String!" },
-    config: { __type: "[CommandKeyValue]" },
+    config: { __type: "[CommandDevicePluginConfiguration]" },
     id: { __type: "ID!" },
     plugin: { __type: "CommandProgramDevicePlugin" },
     rules: { __type: "CommandProgramFlow" },
   },
+  CommandDevicePluginConfiguration: {
+    __typename: { __type: "String!" },
+    id: { __type: "ID" },
+    key: { __type: "CommandProgramDevicePluginConfiguration" },
+    value: { __type: "String" },
+  },
   CommandDevicePluginConfigurationInput: {
+    id: { __type: "ID" },
     key: { __type: "String" },
     value: { __type: "String" },
   },
@@ -416,10 +443,17 @@ export const generatedSchema = {
   },
   CommandDeviceSetpoint: {
     __typename: { __type: "String!" },
+    device: { __type: "CommandProgramDevice" },
     id: { __type: "ID!" },
     key: { __type: "CommandProgramDeviceState" },
     name: { __type: "String" },
     type: { __type: "String" },
+    value: { __type: "String" },
+  },
+  CommandDeviceSetpointCalibration: {
+    __typename: { __type: "String!" },
+    id: { __type: "ID" },
+    setpoint: { __type: "CommandDeviceSetpoint" },
     value: { __type: "String" },
   },
   CommandDeviceSnapshot: {
@@ -617,6 +651,13 @@ export const generatedSchema = {
     name: { __type: "String" },
     trigger: { __type: "String" },
   },
+  CommandProgramDataDeviceInterlockInput: {
+    assertion: { __type: "CommandAssertionInput" },
+    comparator: { __type: "String" },
+    deviceKey: { __type: "String" },
+    inputDevice: { __type: "String" },
+    inputDeviceKey: { __type: "String" },
+  },
   CommandProgramDevice: {
     __typename: { __type: "String!" },
     actions: { __type: "[CommandProgramDeviceAction]" },
@@ -668,6 +709,7 @@ export const generatedSchema = {
   },
   CommandProgramDevicePlaceholder: {
     __typename: { __type: "String!" },
+    dataInterlocks: { __type: "[CommandDataInterlock]" },
     id: { __type: "ID!" },
     interlocks: { __type: "[CommandInterlock]" },
     name: { __type: "String" },
@@ -904,6 +946,14 @@ export const generatedSchema = {
       __type: "CommandProgram!",
       __args: { input: "CommandProgramInput!" },
     },
+    createCommandProgramDataDeviceInterlock: {
+      __type: "CommandDataInterlock!",
+      __args: {
+        device: "ID!",
+        input: "CommandProgramDataDeviceInterlockInput!",
+        program: "ID!",
+      },
+    },
     createCommandProgramDevice: {
       __type: "CommandProgramDevicePlaceholder!",
       __args: { input: "CommandProgramDeviceInput!", program: "ID!" },
@@ -1003,6 +1053,10 @@ export const generatedSchema = {
       __args: { id: "ID" },
     },
     deleteCommandProgram: { __type: "Boolean!", __args: { id: "ID!" } },
+    deleteCommandProgramDataDeviceInterlock: {
+      __type: "Boolean!",
+      __args: { device: "ID!", id: "ID!", program: "ID!" },
+    },
     deleteCommandProgramDevice: {
       __type: "Boolean!",
       __args: { id: "ID!", program: "ID!" },
@@ -1083,6 +1137,10 @@ export const generatedSchema = {
       __type: "[CommandDeviceReport]",
       __args: { device: "ID", grid: "[CommandDeviceReportInput]" },
     },
+    updateCommandDeviceSetpoint: {
+      __type: "String",
+      __args: { device: "ID!", setpoint: "ID!", value: "String" },
+    },
     updateCommandDeviceUptime: {
       __type: "CommandDevice!",
       __args: { uptime: "DateTime", where: "CommandDeviceWhere!" },
@@ -1090,6 +1148,15 @@ export const generatedSchema = {
     updateCommandProgram: {
       __type: "CommandProgram!",
       __args: { id: "ID!", input: "CommandProgramInput!" },
+    },
+    updateCommandProgramDataDeviceInterlock: {
+      __type: "CommandDataInterlock!",
+      __args: {
+        device: "ID!",
+        id: "ID!",
+        input: "CommandProgramDataDeviceInterlockInput!",
+        program: "ID!",
+      },
     },
     updateCommandProgramDevice: {
       __type: "CommandProgramDevicePlaceholder!",
@@ -1266,6 +1333,17 @@ export interface CommandAssertion {
   variable?: Maybe<CommandProgramVariable>;
 }
 
+export interface CommandDataInterlock {
+  __typename?: "CommandDataInterlock";
+  assertion?: Maybe<CommandAssertion>;
+  comparator?: Maybe<ScalarsEnums["String"]>;
+  device?: Maybe<CommandProgramDevicePlaceholder>;
+  deviceKey?: Maybe<CommandProgramDeviceState>;
+  id: ScalarsEnums["ID"];
+  inputDevice?: Maybe<CommandProgramDevicePlaceholder>;
+  inputDeviceKey?: Maybe<CommandProgramDeviceState>;
+}
+
 export interface CommandDevice {
   __typename?: "CommandDevice";
   activeProgram?: Maybe<CommandProgram>;
@@ -1281,6 +1359,7 @@ export interface CommandDevice {
   organisation?: Maybe<HiveOrganisation>;
   peripherals?: Maybe<Array<Maybe<CommandDevicePeripheral>>>;
   reports?: Maybe<Array<Maybe<CommandDeviceReport>>>;
+  setpoints?: Maybe<Array<Maybe<CommandDeviceSetpointCalibration>>>;
   waitingForActions?: Maybe<Array<Maybe<CommandProgramAction>>>;
 }
 
@@ -1322,10 +1401,17 @@ export interface CommandDevicePeripheralProduct {
 
 export interface CommandDevicePlugin {
   __typename?: "CommandDevicePlugin";
-  config?: Maybe<Array<Maybe<CommandKeyValue>>>;
+  config?: Maybe<Array<Maybe<CommandDevicePluginConfiguration>>>;
   id: ScalarsEnums["ID"];
   plugin?: Maybe<CommandProgramDevicePlugin>;
   rules?: Maybe<CommandProgramFlow>;
+}
+
+export interface CommandDevicePluginConfiguration {
+  __typename?: "CommandDevicePluginConfiguration";
+  id?: Maybe<ScalarsEnums["ID"]>;
+  key?: Maybe<CommandProgramDevicePluginConfiguration>;
+  value?: Maybe<ScalarsEnums["String"]>;
 }
 
 export interface CommandDeviceReport {
@@ -1350,10 +1436,18 @@ export interface CommandDeviceReport {
 
 export interface CommandDeviceSetpoint {
   __typename?: "CommandDeviceSetpoint";
+  device?: Maybe<CommandProgramDevice>;
   id: ScalarsEnums["ID"];
   key?: Maybe<CommandProgramDeviceState>;
   name?: Maybe<ScalarsEnums["String"]>;
   type?: Maybe<ScalarsEnums["String"]>;
+  value?: Maybe<ScalarsEnums["String"]>;
+}
+
+export interface CommandDeviceSetpointCalibration {
+  __typename?: "CommandDeviceSetpointCalibration";
+  id?: Maybe<ScalarsEnums["ID"]>;
+  setpoint?: Maybe<CommandDeviceSetpoint>;
   value?: Maybe<ScalarsEnums["String"]>;
 }
 
@@ -1577,6 +1671,7 @@ export interface CommandProgramDeviceConfiguration {
 
 export interface CommandProgramDevicePlaceholder {
   __typename?: "CommandProgramDevicePlaceholder";
+  dataInterlocks?: Maybe<Array<Maybe<CommandDataInterlock>>>;
   id: ScalarsEnums["ID"];
   interlocks?: Maybe<Array<Maybe<CommandInterlock>>>;
   name?: Maybe<ScalarsEnums["String"]>;
@@ -1775,6 +1870,11 @@ export interface Mutation {
   createCommandProgram: (args: {
     input: CommandProgramInput;
   }) => CommandProgram;
+  createCommandProgramDataDeviceInterlock: (args: {
+    device: Scalars["ID"];
+    input: CommandProgramDataDeviceInterlockInput;
+    program: Scalars["ID"];
+  }) => CommandDataInterlock;
   createCommandProgramDevice: (args: {
     input: CommandProgramDeviceInput;
     program: Scalars["ID"];
@@ -1847,6 +1947,11 @@ export interface Mutation {
   }) => Maybe<CommandDeviceReport>;
   deleteCommandProgram: (args: {
     id: Scalars["ID"];
+  }) => ScalarsEnums["Boolean"];
+  deleteCommandProgramDataDeviceInterlock: (args: {
+    device: Scalars["ID"];
+    id: Scalars["ID"];
+    program: Scalars["ID"];
   }) => ScalarsEnums["Boolean"];
   deleteCommandProgramDevice: (args: {
     id: Scalars["ID"];
@@ -1936,6 +2041,11 @@ export interface Mutation {
     device?: Maybe<Scalars["ID"]>;
     grid?: Maybe<Array<Maybe<CommandDeviceReportInput>>>;
   }) => Maybe<Array<Maybe<CommandDeviceReport>>>;
+  updateCommandDeviceSetpoint: (args: {
+    device: Scalars["ID"];
+    setpoint: Scalars["ID"];
+    value?: Maybe<Scalars["String"]>;
+  }) => Maybe<ScalarsEnums["String"]>;
   updateCommandDeviceUptime: (args: {
     uptime?: Maybe<Scalars["DateTime"]>;
     where: CommandDeviceWhere;
@@ -1944,6 +2054,12 @@ export interface Mutation {
     id: Scalars["ID"];
     input: CommandProgramInput;
   }) => CommandProgram;
+  updateCommandProgramDataDeviceInterlock: (args: {
+    device: Scalars["ID"];
+    id: Scalars["ID"];
+    input: CommandProgramDataDeviceInterlockInput;
+    program: Scalars["ID"];
+  }) => CommandDataInterlock;
   updateCommandProgramDevice: (args: {
     id: Scalars["ID"];
     input: CommandProgramDeviceInput;
@@ -2064,14 +2180,17 @@ export interface Subscription {
 export interface SchemaObjectTypes {
   CommandActionItem: CommandActionItem;
   CommandAssertion: CommandAssertion;
+  CommandDataInterlock: CommandDataInterlock;
   CommandDevice: CommandDevice;
   CommandDevicePeripheral: CommandDevicePeripheral;
   CommandDevicePeripheralMap: CommandDevicePeripheralMap;
   CommandDevicePeripheralPort: CommandDevicePeripheralPort;
   CommandDevicePeripheralProduct: CommandDevicePeripheralProduct;
   CommandDevicePlugin: CommandDevicePlugin;
+  CommandDevicePluginConfiguration: CommandDevicePluginConfiguration;
   CommandDeviceReport: CommandDeviceReport;
   CommandDeviceSetpoint: CommandDeviceSetpoint;
+  CommandDeviceSetpointCalibration: CommandDeviceSetpointCalibration;
   CommandDeviceSnapshot: CommandDeviceSnapshot;
   CommandDeviceTimeseriesData: CommandDeviceTimeseriesData;
   CommandDeviceTimeseriesTotal: CommandDeviceTimeseriesTotal;
@@ -2119,14 +2238,17 @@ export interface SchemaObjectTypes {
 export type SchemaObjectTypesNames =
   | "CommandActionItem"
   | "CommandAssertion"
+  | "CommandDataInterlock"
   | "CommandDevice"
   | "CommandDevicePeripheral"
   | "CommandDevicePeripheralMap"
   | "CommandDevicePeripheralPort"
   | "CommandDevicePeripheralProduct"
   | "CommandDevicePlugin"
+  | "CommandDevicePluginConfiguration"
   | "CommandDeviceReport"
   | "CommandDeviceSetpoint"
+  | "CommandDeviceSetpointCalibration"
   | "CommandDeviceSnapshot"
   | "CommandDeviceTimeseriesData"
   | "CommandDeviceTimeseriesTotal"