+++ packages/app/hivecommand-frontend/src/views/Editor/pages/devices/DeviceSingle.tsx
@@ -274,6 +274,34 @@ export const DeviceSingle: React.FC<DeviceSingleProps> = (props) => {
 						}
 						value
 					}
+					dataInterlocks {
+						id
+						inputDevice {
+							id 
+							name
+						}
+						inputDeviceKey { 
+							id
+							key
+						}
+						comparator
+						assertion {
+							type
+							value
+							setpoint {
+								id
+								name
+							}
+							variable {
+								id
+								name
+							}
+						}
+						deviceKey {
+							id 
+							key
+						}
+					}
 					interlocks {
 						id
 	
@@ -345,7 +373,7 @@ export const DeviceSingle: React.FC<DeviceSingleProps> = (props) => {
 	})
 
 	const refetch = () => {
-		return client.refetchQueries({ include: ['Q'] })
+		return client.refetchQueries({ include: ['GetDeviceInfo'] })
 	}
 
 	console.log({data})