+++ packages/core/command-process/src/action/index.ts
@@ -73,7 +73,7 @@ export class ProcessAction {
 			performOperation: this.process.performOperation,
 			getState: this.process.getState,
             setState: this.process.setState,
-            getVariable: this.process.getVariable,
+            valueBank: this.process.valueBank,
 			actions: this.actions
 		}, this.process.getProcess) || {}
 
@@ -163,7 +163,7 @@ export class ProcessAction {
                 performOperation: this.process.performOperation,
                 getState: this.process.getState,
                 setState: this.process.setState,
-                getVariable: this.process.getVariable,
+                valueBank: this.process.valueBank,
                 actions: this.actions
             }, this.process.getProcess)