+++ packages/network/hive-command-discovery-server/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.52](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.51...v1.4.15-alpha.52) (2022-05-27)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.51](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.50...v1.4.15-alpha.51) (2022-05-26)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.50](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.49...v1.4.15-alpha.50) (2022-05-26)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.49](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.48...v1.4.15-alpha.49) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.43](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.42...v1.4.15-alpha.43) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.42](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.41...v1.4.15-alpha.42) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.38](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.37...v1.4.15-alpha.38) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.37](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.36...v1.4.15-alpha.37) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
+## [1.4.15-alpha.36](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.35...v1.4.15-alpha.36) (2022-05-25)
+
+**Note:** Version bump only for package @hive-command/discovery-server
+
+
+
+
+
 ## [1.4.15-alpha.35](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.34...v1.4.15-alpha.35) (2022-05-22)
 
 **Note:** Version bump only for package @hive-command/discovery-server